+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Increase the default socket timeout to 30 seconds.
+
+### Fixed
+- Use the connect timeout value in native socket connect attempts.
 
 ## [2.2.2] - 2019-07-02
 ### Changed